+++ frontend/src/pages/OverviewDash/MostCommonVulns.tsx
