+++ frontend/src/pages/OverviewDash/TopVulnerableDomains.tsx
