+++ frontend/src/pages/OverviewDash/VulnerabilityBarChart.tsx
