+++ frontend/src/pages/Vulnerability/CveSection.tsx
@@ -12,7 +12,7 @@ import {
   TableRow,
   Typography
 } from '@mui/material';
-import { getCVSSColor } from 'pages/Risk/utils';
+import { getCVSSColor } from 'pages/OverviewDash/utils';
 import {
   Cve as CveType,
   Cpe as ProductInfoType,
