+++ frontend/src/pages/Vulnerability/Vulnerability.tsx
@@ -20,7 +20,7 @@ import {
   Typography
 } from '@mui/material';
 import { tableCellClasses } from '@mui/material/TableCell';
-import { getSeverityColor } from 'pages/Risk/utils';
+import { getSeverityColor } from 'pages/OverviewDash/utils';
 import { useAuthContext } from 'context';
 import {
   Cve as CveType,
